feat OpenAI API 개발, 프롬프터 추천 기능 개발
- vectorDB 설정 추가
 - converter 개발
 - service 기능 추가

chore vectorDB 및 openAI 설정 추가

+++ src/main/java/ai/herofactoryservice/create_game_resource_service/service/OpenAiApi.java
@@ -81,7 +81,7 @@ public class OpenAiApi {
         }
     }
 
-    public Float[] embeddings(String text) {
+    public double[] embeddings(String text) {
         String url = baseUrl + "/embeddings";
 
         try {
@@ -106,9 +106,9 @@ public class OpenAiApi {
                         .get(0)
                         .path("embedding");
 
-                Float[] embeddings = new Float[embeddingData.size()];
+                double[] embeddings = new double[embeddingData.size()];
                 for (int i = 0; i < embeddingData.size(); i++) {
-                    embeddings[i] = (float) embeddingData.get(i).asDouble();
+                    embeddings[i] = embeddingData.get(i).asDouble();
                 }
                 return embeddings;
             }
